Refactor, fix formatting
