+++ hydrate/cluster.py
@@ -2,6 +2,7 @@
 from kubernetes import client, config
 from .component import Component
 
+
 class Cluster():
     """Define Cluster data and methods."""
